# Fix and optimize try_cast