Summary:
try_cast is translated before this diff to try(cast()), hence it would hide errors
that happen during the evaluation of the input expression of the cast which is wrong.

This diff fixes that bug by making sure that try_cast does not hide input
expression errors, in addition to that. The diff optimizes casting empty strings
to primitives by completely avoiding the throw.

try_cast_invalid_empty_input went down form around 500ms to 1.85ms .
```
============================================================================
[...]hmarks/ExpressionBenchmarkBuilder.cpp     relative  time/iter   iters/s
============================================================================
cast_int##try_cast_invalid_empty_input                      1.85ms    539.56
cast_int##tryexpr_cast_invalid_empty_input                471.92ms      2.12
cast_int##try_cast_invalid_nan                               1.14s   876.49m
cast_int##tryexpr_cast_invalid_nan                           1.94s   516.60m
cast_int##try_cast_valid                                    2.70ms    370.73
cast_int##tryexpr_cast_valid                                2.66ms    375.64
cast_int##cast_valid                                        2.48ms    403.00
```

Differential Revision: D47439839